change update message style
