+++ packages/api/src/service/discord.ts
@@ -1,4 +1,12 @@
-import { Client, Guild } from 'discord.js'
+import ansi from 'ansi-escape-sequences'
+import {
+  AttachmentPayload,
+  Client,
+  Guild,
+  MessageCreateOptions,
+  MessageEditOptions,
+} from 'discord.js'
+import { Episode, MediaInfo, Season } from '../model/Media'
 
 import { config } from '../config'
 
@@ -29,4 +37,138 @@ export function getTextChannel(server: Guild) {
         cause: error,
       })
     })
+}
+
+type StatusMeta = {
+  /** hex rgb number */
+  color: number
+  /** png image 1x1 pixel */
+  base64: string
+  ansi: keyof typeof ansi.style
+}
+
+function colorFromStatus(status: MediaInfo['status']): StatusMeta {
+  switch (status) {
+    case 'Available':
+      return {
+        color: 0x2ecc71,
+        base64:
+          'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAA1JREFUGFdj0DtT+B8ABQICa3znZ2oAAAAASUVORK5CYII=',
+        ansi: 'green',
+      }
+    case 'Blacklisted':
+      return {
+        color: 0xe74c3c,
+        base64:
+          'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAA1JREFUGFdjeO5j8x8ABfwCb6cFtH0AAAAASUVORK5CYII=',
+        ansi: 'red',
+      }
+    case 'Pending':
+      return {
+        color: 0xe67e22,
+        base64:
+          'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAA1JREFUGFdjeFan9B8ABloChnlkdVsAAAAASUVORK5CYII=',
+        ansi: 'yellow',
+      }
+    case 'Processing':
+      return {
+        color: 0x3498db,
+        base64:
+          'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAA1JREFUGFdjMJlx+z8ABVICp/IuGtoAAAAASUVORK5CYII=',
+        ansi: 'blue',
+      }
+    default:
+      return {
+        color: 0x95a5a6,
+        base64:
+          'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAAAXNSR0IArs4c6QAAAA1JREFUGFdjmLp02X8ABpMC4IYnRY4AAAAASUVORK5CYII=',
+        ansi: 'white',
+      }
+  }
+}
+
+function statusMessageTemplate(
+  status: MediaInfo['status'],
+  text: string,
+): MessageCreateOptions {
+  return {
+    embeds: [
+      {
+        title: text,
+        color: colorFromStatus(status).color,
+      },
+    ],
+  }
+}
+
+export const templates = {
+  mediaStatus: function (media: MediaInfo): MessageCreateOptions {
+    return statusMessageTemplate(media.status, `Status → ${media.status}`)
+  },
+
+  episodeStatus: function (episode: Episode): MessageCreateOptions {
+    return statusMessageTemplate(
+      episode.available ? 'Available' : 'Unknown',
+      `Episode S${episode.seasonNumber}E${episode.number} → ${
+        episode.available ? 'Available' : 'Unavailable'
+      }`,
+    )
+  },
+
+  seasonStatus: function (season: Season): MessageCreateOptions {
+    return statusMessageTemplate(
+      season.available ? 'Available' : 'Unknown',
+      `Season ${season.number} → ${
+        season.available ? 'Available' : 'Unavailable'
+      }`,
+    )
+  },
+
+  mainMessage: function (
+    media: MediaInfo,
+  ): MessageCreateOptions & MessageEditOptions {
+    const requests = media.requests
+      .map((r) => (r.user.discordId ? `<@${r.user.discordId}>` : r.user.name))
+      .join(' ')
+    const statusText =
+      media.status === 'Processing'
+        ? `Processing — ${(media.downloadStatus.completion * 100).toFixed()}%`
+        : media.status
+
+    return {
+      embeds: [
+        {
+          title: media.title,
+          url: media.link,
+          description: media.overview,
+          thumbnail: {
+            url: media.image,
+          },
+          color: colorFromStatus(media.status).color,
+          fields: requests
+            ? [
+                {
+                  name: '',
+                  value: `*Requested by* ${requests}`,
+                  inline: true,
+                },
+              ]
+            : undefined,
+          footer: {
+            text: statusText,
+            icon_url: 'attachment://status',
+          },
+        },
+      ],
+      files: [
+        {
+          name: 'status',
+          attachment: Buffer.from(
+            colorFromStatus(media.status).base64,
+            'base64',
+          ),
+        } satisfies AttachmentPayload,
+      ],
+    }
+  },
 }
